+++ egui_demo_lib/src/lib.rs
@@ -148,8 +148,8 @@ fn test_egui_e2e() {
         let full_output = ctx.run(raw_input.clone(), |ctx| {
             demo_windows.ui(ctx);
         });
-        let clipped_meshes = ctx.tessellate(full_output.shapes);
-        assert!(!clipped_meshes.is_empty());
+        let clipped_primitives = ctx.tessellate(full_output.shapes);
+        assert!(!clipped_primitives.is_empty());
     }
 }
 
@@ -167,8 +167,11 @@ fn test_egui_zero_window_size() {
         let full_output = ctx.run(raw_input.clone(), |ctx| {
             demo_windows.ui(ctx);
         });
-        let clipped_meshes = ctx.tessellate(full_output.shapes);
-        assert!(clipped_meshes.is_empty(), "There should be nothing to show");
+        let clipped_primitives = ctx.tessellate(full_output.shapes);
+        assert!(
+            clipped_primitives.is_empty(),
+            "There should be nothing to show"
+        );
     }
 }